llm/requirements.txt
openai~=1.77.0
python-dotenv~=1.1.0
google~=3.0.0
